Refactored name_disambiguation
Person and PeopleDatabase now each have their own files
