+++ data/name_disambiguation/clean_org_names_to_raw_org_names.json
@@ -174,10 +174,14 @@
     "Inbifo": [
         "INBIFO, INSTITUT FUR BIOLOGISCHE FORSCHUNG"
     ],
+    "Jacob & Medinger": [
+        "JACOB AND MEDINGER"
+    ],
     "Johns Hopkins University": [
         "JOHNS HOPKINS UNIV",
         "JOHNS HOPKINS UNIV SCHOOL MEDICINE",
-        "JOHNS HOPKINS UNIV SCHOOL OF MEDICINE"
+        "JOHNS HOPKINS UNIV SCHOOL OF MEDICINE",
+        "JOHNS HOPKINS SCHOOL OF HYGIENE"
     ],
     "Jones, Day, Reavis & Pogue": [
         "JONES DAY",
